requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
alembic==1.14.1
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.8.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorlog==6.9.0
contourpy==1.1.1
cycler==0.12.1
docopt==0.6.2
einops==0.8.1
ffmpeg==1.4
filelock==3.16.1
flask==3.0.3
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
greenlet==3.1.1
huggingface-hub==0.30.2
HyperPyYAML==1.2.2
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.7
lightning==2.3.3
lightning-utilities==0.11.9
llvmlite==0.41.1
mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai-whisper==20240930
optuna==4.3.0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
primePy==1.3
propcache==0.2.0
protobuf==5.29.4
pyannote.audio==3.3.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.8
scikit-learn==1.3.2
scipy==1.10.1
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
sqlalchemy==2.0.40
sympy==1.13.3
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.7.0
torch==2.4.1
torch-audiomentations==0.12.0
torch-pitch-shift==1.2.5
torchaudio==2.4.1
torchmetrics==1.5.2
tqdm==4.67.1
triton==3.0.0
typer==0.15.3
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
werkzeug==3.0.6
yarl==1.15.2
zipp==3.20.2
